Windows port: ttnn_core first-wave MSVC fixes
- operation.hpp: include <experimental/type_traits> only where it exists
- serialization / overlapped_serialization: mmap-based tensor loads read the
  file into memory on Windows (no mmap); POSIX fd/stat path kept elsewhere
- tensor_impl / h2d_socket_service: POSIX includes behind _WIN32 guards;
  page size via GetSystemInfo
- graph_processor: std::array iterators are not raw pointers on MSVC's STL
- explicit narrowing casts in braced initializers (reports, graph_trace_utils,
  tensor_spec_flatbuffer, tensor_utils, xtensor/partition) -- MSVC hard-errors
  where GNU/Clang builds run with -Wno-narrowing

+++ ttnn/core/graph/graph_trace_utils.cpp
@@ -268,7 +268,7 @@ std::vector<TensorInfo> extract_output_info(const nlohmann::json& trace) {
             const std::string shape_string = tensor[kParams][kShape];
             const auto shape = parse_shape(shape_string);
 
-            output.emplace_back(TensorInfo{.shape = shape, .size = size, .type = type});
+            output.emplace_back(TensorInfo{.shape = shape, .size = static_cast<uint32_t>(size), .type = type});
         }
     }
